requirements.txt
celery==5.2.7
click==8.1.3
Django==4.1.7
djangorestframework==3.14.0
environs==9.5.0
flake8==6.0.0
Pyrogram==2.0.100
PySocks==1.7.1
pytube==12.1.2
requests==2.28.2
TgCrypto==1.2.5
gunicorn==20.1.0
psycopg2-binary==2.9.5